UI-split-config-rendering-modules-for-CI
