+++ ui/src/ui/views/config-form.node.ts
@@ -1,6 +1,15 @@
-import { html, nothing, type TemplateResult } from "lit";
 import JSON5 from "json5";
+import { html, nothing, type TemplateResult } from "lit";
 import type { ConfigUiHints } from "../types.ts";
+import {
+  buildHelpText,
+  icons,
+  isAnySchema,
+  jsonValue,
+  renderFieldHeader,
+  renderHelpPopover,
+  renderRawFallback,
+} from "./config-form.node.helpers.ts";
 import {
   defaultValue,
   hintForPath,
@@ -10,253 +19,6 @@ import {
   schemaType,
   type JsonSchema,
 } from "./config-form.shared.ts";
-
-const META_KEYS = new Set(["title", "description", "default", "nullable"]);
-
-function isAnySchema(schema: JsonSchema): boolean {
-  const keys = Object.keys(schema ?? {}).filter((key) => !META_KEYS.has(key));
-  return keys.length === 0;
-}
-
-function jsonValue(value: unknown): string {
-  if (value === undefined) {
-    return "";
-  }
-  try {
-    return JSON.stringify(value, null, 2) ?? "";
-  } catch {
-    return "";
-  }
-}
-
-function shortValue(value: unknown, max = 80): string {
-  const raw = jsonValue(value).replace(/\s+/g, " ").trim();
-  if (!raw) {
-    return "";
-  }
-  if (raw.length <= max) {
-    return raw;
-  }
-  return `${raw.slice(0, max - 3)}...`;
-}
-
-function fallbackHelpForType(type: string | undefined): string {
-  switch (type) {
-    case "string":
-      return "Enter a text value.";
-    case "number":
-    case "integer":
-      return "Enter a numeric value.";
-    case "boolean":
-      return "Toggle this setting on or off.";
-    case "array":
-      return "Manage one or more values.";
-    case "object":
-      return "Configure grouped settings.";
-    default:
-      return "Configure this setting.";
-  }
-}
-
-function buildHelpText(schema: JsonSchema, explicitHelp?: string): string {
-  const help = explicitHelp?.trim() || schema.description?.trim() || "";
-  const parts: string[] = [];
-  if (help) {
-    parts.push(help);
-  } else {
-    parts.push(fallbackHelpForType(schemaType(schema)));
-  }
-  if (schema.enum && schema.enum.length > 0) {
-    parts.push(
-      `Allowed values: ${schema.enum
-        .map((entry) => shortValue(entry, 24))
-        .filter(Boolean)
-        .join(", ")}.`,
-    );
-  }
-  if (schema.default !== undefined) {
-    const def = shortValue(schema.default, 60);
-    if (def) {
-      parts.push(`Default: ${def}.`);
-    }
-  }
-  return parts.join(" ");
-}
-
-function positionHelpPopover(details: HTMLDetailsElement): void {
-  const doc = details.ownerDocument;
-  const view = doc.defaultView;
-  const panel = details.querySelector<HTMLElement>(".cfg-help__panel");
-  if (!view || !panel) {
-    return;
-  }
-
-  const viewportPadding = 12;
-  const maxWidth = Math.max(180, Math.min(360, view.innerWidth - viewportPadding * 2));
-  details.style.setProperty("--cfg-help-max-width", `${maxWidth}px`);
-
-  const triggerRect = details.getBoundingClientRect();
-  const panelWidth = Math.min(maxWidth, panel.scrollWidth || maxWidth);
-  const leftIfRightAligned = triggerRect.right - panelWidth;
-  const rightIfLeftAligned = triggerRect.left + panelWidth;
-
-  let align: "right" | "left" = "right";
-  if (leftIfRightAligned < viewportPadding && rightIfLeftAligned <= view.innerWidth - viewportPadding) {
-    align = "left";
-  } else if (leftIfRightAligned < viewportPadding && rightIfLeftAligned > view.innerWidth - viewportPadding) {
-    const spaceOnLeft = triggerRect.right;
-    const spaceOnRight = view.innerWidth - triggerRect.left;
-    align = spaceOnRight >= spaceOnLeft ? "left" : "right";
-  }
-
-  details.dataset.align = align;
-}
-
-function renderHelpPopover(label: string, helpText: string): TemplateResult {
-  return html`
-    <details
-      class="cfg-help"
-      @toggle=${(event: Event) => {
-        const details = event.currentTarget as HTMLDetailsElement;
-        if (!details.open) {
-          delete details.dataset.align;
-          return;
-        }
-        const doc = details.ownerDocument;
-        doc.querySelectorAll<HTMLDetailsElement>(".cfg-help[open]").forEach((entry) => {
-          if (entry !== details) {
-            entry.open = false;
-          }
-        });
-        const alignPopover = () => positionHelpPopover(details);
-        alignPopover();
-        doc.defaultView?.requestAnimationFrame(alignPopover);
-      }}
-    >
-      <summary
-        class="cfg-help__trigger"
-        aria-label=${`Help for ${label}`}
-        @click=${(event: Event) => event.stopPropagation()}
-      >
-        ?
-      </summary>
-      <div class="cfg-help__panel">${helpText}</div>
-    </details>
-  `;
-}
-
-function renderFieldHeader(
-  label: string,
-  helpText: string,
-  showLabel: boolean,
-  showHelp = true,
-): TemplateResult | typeof nothing {
-  if (!showLabel) {
-    return nothing;
-  }
-  if (!showHelp) {
-    return html`
-      <div class="cfg-field__header">
-        <label class="cfg-field__label">${label}</label>
-      </div>
-    `;
-  }
-  return html`
-    <div class="cfg-field__header">
-      <label class="cfg-field__label">${label}</label>
-      ${renderHelpPopover(label, helpText)}
-    </div>
-  `;
-}
-
-function renderRawFallback(params: {
-  schema: JsonSchema;
-  value: unknown;
-  path: Array<string | number>;
-  hints: ConfigUiHints;
-  disabled: boolean;
-  reason: string;
-  showLabel?: boolean;
-  onPatch: (path: Array<string | number>, value: unknown) => void;
-}): TemplateResult {
-  const { schema, value, path, hints, disabled, reason, onPatch } = params;
-  const showLabel = params.showLabel ?? true;
-  const hint = hintForPath(path, hints);
-  const label = hint?.label ?? schema.title ?? humanize(String(path.at(-1)));
-  const helpText = buildHelpText(schema, hint?.help);
-  const fallback = jsonValue(value);
-  return html`
-    <div class="cfg-field cfg-field--fallback cfg-item cfg-item--full">
-      ${renderFieldHeader(label, helpText, showLabel)}
-      <div class="cfg-field__warning">${reason}</div>
-      <textarea
-        class="cfg-textarea cfg-textarea--sm"
-        placeholder="JSON5 value"
-        rows="4"
-        .value=${fallback}
-        ?disabled=${disabled}
-        @change=${(event: Event) => {
-          const target = event.target as HTMLTextAreaElement;
-          const raw = target.value.trim();
-          if (!raw) {
-            onPatch(path, undefined);
-            target.setCustomValidity("");
-            return;
-          }
-          try {
-            onPatch(path, JSON5.parse(raw));
-            target.setCustomValidity("");
-          } catch {
-            target.setCustomValidity("Invalid JSON5 value");
-            target.reportValidity();
-            target.value = fallback;
-          }
-        }}
-      ></textarea>
-    </div>
-  `;
-}
-
-const icons = {
-  chevronDown: html`
-    <svg
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    >
-      <polyline points="6 9 12 15 18 9"></polyline>
-    </svg>
-  `,
-  plus: html`
-    <svg
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    >
-      <line x1="12" y1="5" x2="12" y2="19"></line>
-      <line x1="5" y1="12" x2="19" y2="12"></line>
-    </svg>
-  `,
-  trash: html`
-    <svg
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      stroke-width="2"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    >
-      <polyline points="3 6 5 6 21 6"></polyline>
-      <path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"></path>
-    </svg>
-  `,
-};
 
 export function renderNode(params: {
   schema: JsonSchema;
@@ -292,7 +54,8 @@ export function renderNode(params: {
   if (schema.anyOf || schema.oneOf) {
     const variants = schema.anyOf ?? schema.oneOf ?? [];
     const nonNull = variants.filter(
-      (entry) => !(entry.type === "null" || (Array.isArray(entry.type) && entry.type.includes("null"))),
+      (entry) =>
+        !(entry.type === "null" || (Array.isArray(entry.type) && entry.type.includes("null"))),
     );
 
     if (nonNull.length === 1) {
@@ -343,8 +106,14 @@ export function renderNode(params: {
     }
 
     const primitiveTypes = new Set(nonNull.map((entry) => schemaType(entry)).filter(Boolean));
-    const normalizedTypes = new Set([...primitiveTypes].map((entry) => (entry === "integer" ? "number" : entry)));
-    if ([...normalizedTypes].every((entry) => ["string", "number", "boolean"].includes(entry as string))) {
+    const normalizedTypes = new Set(
+      [...primitiveTypes].map((entry) => (entry === "integer" ? "number" : entry)),
+    );
+    if (
+      [...normalizedTypes].every((entry) =>
+        ["string", "number", "boolean"].includes(entry as string),
+      )
+    ) {
       const hasString = normalizedTypes.has("string");
       const hasNumber = normalizedTypes.has("number");
       const hasBoolean = normalizedTypes.has("boolean");
@@ -389,7 +158,9 @@ export function renderNode(params: {
                 <button
                   type="button"
                   class="cfg-segmented__btn ${
-                    option === resolvedValue || String(option) === String(resolvedValue) ? "active" : ""
+                    option === resolvedValue || String(option) === String(resolvedValue)
+                      ? "active"
+                      : ""
                   }"
                   ?disabled=${disabled}
                   @click=${() => onPatch(path, option)}
@@ -478,7 +249,11 @@ function renderTextInput(params: {
   const isSensitive = hint?.sensitive ?? isSensitivePath(path);
   const placeholder =
     hint?.placeholder ??
-    (isSensitive ? "******" : schema.default !== undefined ? `Default: ${String(schema.default)}` : "");
+    (isSensitive
+      ? "******"
+      : schema.default !== undefined
+        ? `Default: ${String(schema.default)}`
+        : "");
   const displayValue = value ?? "";
 
   return html`
@@ -548,7 +323,9 @@ function renderNumberInput(params: {
   const helpText = buildHelpText(schema, hint?.help);
   const displayValue = value ?? schema.default ?? "";
   const numValue = typeof displayValue === "number" ? displayValue : 0;
-  const numberFieldClass = showLabel ? "cfg-field--number" : "cfg-field--number cfg-field--number-no-label";
+  const numberFieldClass = showLabel
+    ? "cfg-field--number"
+    : "cfg-field--number cfg-field--number-no-label";
 
   return html`
     <div class="cfg-field cfg-field--compact ${numberFieldClass} cfg-item cfg-item--compact">
@@ -673,7 +450,12 @@ function renderObject(params: {
         : null;
   const primitiveChildCount = sorted.filter(([, node]) => {
     const nodeType = schemaType(node);
-    return nodeType === "string" || nodeType === "number" || nodeType === "integer" || nodeType === "boolean";
+    return (
+      nodeType === "string" ||
+      nodeType === "number" ||
+      nodeType === "integer" ||
+      nodeType === "boolean"
+    );
   }).length;
   const complexChildCount = sorted.length - primitiveChildCount + (mapSchema ? 1 : 0);
   const fullWidth = complexChildCount > 4 || sorted.length >= 12;
@@ -816,7 +598,8 @@ function renderArray(params: {
   const arr = Array.isArray(value) ? value : Array.isArray(schema.default) ? schema.default : [];
   const itemType = schemaType(itemsSchema);
   const hasComplexItems = arr.some((entry) => typeof entry === "object" && entry !== null);
-  const fullWidth = itemType === "object" || itemType === "array" || hasComplexItems || arr.length >= 4;
+  const fullWidth =
+    itemType === "object" || itemType === "array" || hasComplexItems || arr.length >= 4;
 
   return html`
     <div class="cfg-array cfg-item ${fullWidth ? "cfg-item--full" : "cfg-item--compact"}">
@@ -909,7 +692,9 @@ function renderMapField(params: {
   const entries = Object.entries(value ?? {}).filter(([key]) => !reservedKeys.has(key));
   const helpText = "Extra key/value entries not explicitly listed in the schema.";
   const valueType = schemaType(schema);
-  const hasComplexEntries = entries.some(([, entryValue]) => typeof entryValue === "object" && entryValue !== null);
+  const hasComplexEntries = entries.some(
+    ([, entryValue]) => typeof entryValue === "object" && entryValue !== null,
+  );
   const schemaIsComplex = !anySchema && (valueType === "object" || valueType === "array");
   const fullWidth = schemaIsComplex || hasComplexEntries || entries.length >= 4;
